Basic implementation of initialize_empty and save_to_file functions
Implemented initialize_empty and save_to_file functions. The latter needs some work (need to fix overwriting functions) but both are functional. Testing code is in the TEST_LOAD menu option, and creates two new .oad files (one is an "initialized" document, and one is a copy of testtext.oad).